third_party/blink/web_tests/platform/generic/fast/forms/time/time-stepup-stepdown-expected.txt

Check stepUp() and stepDown() behavior for type=time.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Invalid value
PASS stepUp("", null, null) is "00:01"
PASS stepDown("", null, null) is "00:00"
Non-number arguments
PASS stepUp("20:13", null, null, "0") is "20:13"
PASS stepDown("20:13", null, null, "0") is "20:13"
PASS stepUp("20:13", null, null, "foo") is "20:13"
PASS stepDown("20:13", null, null, "foo") is "20:13"
PASS stepUp("20:13", null, null, null) is "20:13"
PASS stepDown("20:13", null, null, null) is "20:13"
Normal cases
PASS stepUp("20:13", null, null) is "20:14"
PASS stepDown("20:13", null, null) is "20:12"
PASS stepUp("20:13", null, null, 10) is "20:23"
PASS stepDown("20:13", null, null, 11) is "20:02"
PASS stepUp("20:13", "4", null, 2) is "20:13:08"
PASS stepDown("20:13", "4", null, 3) is "20:12:48"
Step=any
PASS stepUp("20:13", "any", null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element does not have an allowed value step..
PASS stepDown("20:13", "any", null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element does not have an allowed value step..
Overflow/underflow
FAIL stepUp("20:13", "3.40282346e+38", null) should be 20:13. Was 00:00:00.
FAIL stepDown("20:13", "3.40282346e+38", null) should be 20:13. Was 00:00:00.
PASS stepUp("20:13", "1", "20:13") is "20:13:00"
PASS stepDown("20:13", "1", "20:13") is "20:13:00"
PASS stepUp("23:59", null, null) is "23:59"
PASS stepDown("00:00", null, null) is "00:00"


Invalid min/max:
PASS setInputAttributes("10:00", "09:00", "60", "08:00"); input.stepUp(); input.value is "08:00"
PASS inputWithInitialValue.stepUp(); inputWithInitialValue.value is "12:34"


Step-mismatched initial values
PASS stepUp("20:13", "86400.000", null, 65536) is "20:13"
PASS stepUp("20:13", "86400.001", null, 65536) is "20:13"
PASS stepUp("20:13", "600", null, 1) is "20:20"
PASS stepUp("20:13", "600", null, 2) is "20:30"
PASS stepDown("20:13", "600", null, 1) is "20:10"
PASS stepDown("20:13", "600", null, 2) is "20:00"


PASS successfullyParsed is true

TEST COMPLETE
